refactor: follow the MTV framework to refacter
