+++ flask101/app/forms.py
@@ -1,5 +1,5 @@
-from flask_wtf import  FlaskForm
-from wtforms import  StringField, PasswordField, BooleanField, SubmitField
+from flask_wtf import FlaskForm
+from wtforms import StringField, PasswordField, BooleanField, SubmitField
 from wtforms.validators import DataRequired
