+++ flask101/app/routes.py
@@ -1,0 +1,31 @@
+from flask import render_template, flash, redirect, url_for
+from app import app
+from app.forms import LoginForm
+
+
+@app.route('/')
+@app.route('/index')
+def index():
+    user = {'username': 'Ally'}
+    posts = [
+        {
+            'author': {'username': 'Kenny'},
+            'body': "Japan's SAKURA is beautiful!"
+        },
+        {
+            'author': {'username': 'Mandy'},
+            'body': "My baby is quite, right?"
+        }
+    ]
+    return render_template(template_name_or_list='index.html', title='Home', user=user, posts=posts)
+
+
+@app.route('/login', methods=['GET', 'POST'])
+def login():
+    form = LoginForm()
+    if form.validate_on_submit():
+        flash("User's {} Login request, Remember me = {}".format(
+            form.username.data, form.remember_me.data
+        ))
+        return redirect(url_for('index'))
+    return render_template(template_name_or_list='login.html', title='Login', form=form)
