+++ flask101/app/templates/base.html
@@ -14,8 +14,8 @@
 <body>
     <div>
         My Blog:
-        <a href="/index"> Home Page</a>
-        <a href="/login"> Login</a>
+        <a href="{{ url_for('index') }}"> Home Page</a>
+        <a href="{{ url_for('login') }}"> Login</a>
     </div>
     <hr>
 <!-- If submit successful-->
