+++ flask101/config.py
@@ -1,0 +1,6 @@
+import os
+
+
+class Config:
+    SECRET_KEY = os.environ.get('SECRET_KEY') or 'you-will-never-guess'
+    SQLALCHEMY_DATABASE_URI = "mysql+pymysql://root:root@localhost:3306/MicroBlogData"
